api/requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
aiohttp==3.12.15

